Add CommitList component and its Flux logic
CommitList displays the commit history of a repo/branch. The
component will fire a GET_COMMITS action if the commit list
of the repo/branch given through props is empty.

The RepoStore responds by calling the GitHubAPI, which fires a
GOT_COMMITS action that prompts the RepoStore
  - to add the new commit objects to the repo's `objs` store, and
  - to add their sha to the commits array of the repo/branch.

For now, no further interactivity is implemented.