Replace GET_COMMITS Action with RepoStore method
Using an action to tell the store to get commits from the GitHub
API caused a problem of Action chaining in the store.

( GET_COMMITS -> API call -> GOT_COMMITS -> ... )

This was raising an error as the Dispatcher refuses new action
calls from inside the store's switch/case block, but it was
failing silently within the promise chain due to lack of a
.done() call at the end.

The error caused the component to fail to load until after a
page refresh.

This emphasizes that Actions should be considered only as
'setters', and any interactions with the store that don't
directly change the state should be done through public
methods instead.