Diff-based styling of file text
After escaping HTML, each line of file text is wrapped in a span
with class of either `changed` or `unchanged`, where `changed`
lines are those marked with a + in the diff patch.

Changed lines are given full opacity and bold font weight, and
unchanged lines are given half opacity and normal font weight,
which doesn't interfere with the normal syntax highlighting.