Move expanded state of FileTree into ViewStore
Moving to Store makes use of already-in-place caching logic so the
expanded state of FileTree is retained through refresh.

Trees that are unchanged across commits will remain expanded when
checkedOut commit changes. However, trees that do change will have
a different sha and so will collapse.

It may be preferable to have trees with changed files automatically
expand, but I'll leave that for now to a separate commit.