Moved GitHub API calls to server side
Client-side API calls become problematic because GitHub requires
authentication for more than a handful of requests per hour, and
auth credentials can't be stored in the client application.

The GitHubAPI file on the client is replaced by the http-handlers
file on the server (which contains most of the logic) and the
ServerAPI file on the client (which is just an ajax interface to
the server).

For now the auth still uses my own GitHub login, but should be
replaced with an OAuth app token.