Style everything in initial commit as `changed`
Initial commit was showing everything as unchanged because there
was no diff to be made with prior commit, and the default style
was `unchanged` unless a patch said otherwise.

A special case was made by adding `commit.diffed = 'all'` to the
first commit and using it to conditionally mark all files/lines
as `changed`.