Load _ and hljs by CDN, fix Promises, minify js
Size of bundle.js was large, in part because of the size of hljs.
The node highlight.js module includes many less-commonly used
languages presumably because it's intended to only be used on the
server. Moving lodash also saves significant size in bundle.js
which makes page render faster. The node modules are still needed
to define globals in unit tests, but are moved to devDependencies.

Using promisify-node on client broke upon minification. A more
standard manual promise implementation of is now used in ServerAPI.

Minification of js added to Gruntfile.

Index.html updated to use bundle.js on localhost and bundle.min.js
elsewhere, and load livereload only on localhost.