Display first line of commit message as header
Simply prepends '# ' to the markdown. Font sizes given for `h1` and
`h2` elements, smaller headers should not occur since commits treat
'#' as comments, so only `====` and `----` headers are expected in
the message body.

Note that the second line of multiline commits should always be
newline, so the first line should not be followed by `====` in any
case.