Prevent highlight.js opacity reduction
When languages are nested as in JSX, the nested language is given
0.5 opacity by default, which combines with the diff styling to
become almost unreadable. Simplest fix is just to override this
behavior. Since the languages retain different colors anyway,
it's still easy to distinguish the transitions.

Note that JSX often seems to break proper highlighting, but
presumably will be fixed in a future update.