Setup getCommits to handle GitHub API pagination
GitHub API getCommits returns up to 30 per page by default, 100
max, in reverse chronological order. To get back to the first
commit for large history, all pages must be stepped through.

This is achieved by dynamically building a promise chain until
the last page is reached, at which point the full history can
be replied to the client. `per_page` is set to 100 to minimize
the total number of requests.

The `github-api` module does not provide pagination metadata
in its getCommits result, so it is replaced with `github`.

Note that repos with many commits may now take a very long time
to load, so it would be a good idea to add a spinner.

The API module change also basically breaks jest, taking ~ 2 min
to run even if all but `require` lines are commented out. Given
the many issues with jest, it's time to switch to a different
testing system. For now, simply commenting out http-handlers-test.