Autoexpand folders containing changes
Since dispatcher complains about action chains, autoexpanding of
folders was implemented in ViewStore as a response to existing
`GOT_DIFF` and `CHECKOUT` actions, as well as `GOT_TREE` if it
is the first commit (when `checkedOut` is 0).

Because this required walking the tree in the same manner that was
already implemented in FileTree for building the JSX, the basic
walk functionality was factored out into `utils/WalkerFactory`,
and more specific logic can be introduced locally by the `onBlob`
and `onTree` methods.

The property `commit.diffed` was also renamed to `commit.changed`
to be more clear and consistent with the CSS class name.