Limit commits to prevent crash with large repos
Since GitHub API only returns commits in reverse chronological
order, reaching the first commit seems to require going through
all the commits in the repo with multiple requests (up to 100
commits per request).

For large repos with thousands of commits, this becomes untenable.
Until an API option for chronological commits is introduced, there
seems to be no good way to handle such repos. For now, only the
most recent commits up to the limit are displayed, along with a
warning message to indicate that the full project isn't displayed.

This is only a problem when using this as a generic viewer of
arbitrary repos, but if applied specifically for tutorials it
shouldn't be an issue since tutorials would contain a manageable
number of steps.